requirements.txt
altair==6.1.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.3
dill==0.4.1
distro==1.9.0
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.20.0
numpy==2.2.6
openai==2.32.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1776209387/work
pandas==2.3.3
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
psutil==7.2.2
pyarrow==24.0.0
pydantic==2.13.2
pydantic_core==2.46.2
pydeck==0.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
qiskit==2.4.0
qiskit-aer==0.17.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
rustworkx==0.17.1
scipy==1.15.3
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
stevedore==5.7.0
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
usd-core==26.3
watchdog==6.0.0
